requirements.txt
flask==2.3.2
groq
beautifulsoup4
request